requirements.txt
django
djangorestframework
firebase_admin